Move property size VS total area charts to trends
